isl_multi_templ.c: extract out isl_multi_identity_templ.c
This makes the function isl_multi_*_identity opt-in rather than opt-out.
